// YerbowoFrontend/src/app/products/models/productDetail.ts
import { ProductState } from '../../shared/enums/productState.enum';

export interface ProductDetail {
   id: number;
   code: string;
   name: string;
   description: string;
   price: number;
   oldPrice: number;
   stock: number;
   state: ProductState;
   image: string;
   category: string;
   subcategory: string;
}