// YerbowoFrontend/src/app/products/models/productCard.ts
import { ProductState } from '../../shared/enums/productState.enum';

export interface ProductCard {
   id: number;
   name: string;
   categorySlug: string;
   subcategorySlug: string;
   price: number;
   oldPrice: number;
   state: ProductState;
   image: string;
   slug: string;
   createdAt: Date;
}